LOG-5798: Refactor functional-benchmarker to obs
